Relocate tensor-type-hint detection from `Function` to `Parameter`.
`Function.nodeIsTensorTypeHint`, the static `getAllAttributes` helper, and the `TF_TENSOR_FQN` constant existed on `Function` only as machinery for `Parameter.hasTensorTypeHint`. Move them onto `Parameter`. `Parameter.hasTensorTypeHint` now does the full AST walk plus FQN-resolution loop inline, reaching `this.function.X()` for the project-context state it still needs (`getContainingDocument`, `getContainingModuleName`, `getContainingFile`, `getNature`, `getProject`). `getAllAttributes` is a private static on `Parameter`; `TF_TENSOR_FQN` and a `Parameter`-scoped `LOG` follow. Spotless picks up the now-unused `Attribute` and `VisitorBase` imports from `Function`.

Also fixes the duplicate-Javadoc glitch around `hasTensorContainer`/`getTensorTypes` introduced in the earlier encapsulation #2 commit (`c0a74f4`): the `getTensorTypes` Javadoc was orphaned above `hasTensorContainer` and `getTensorTypes` itself had none. Each method now has its own Javadoc immediately preceding it.
